Added 2 new variables to Task, isOverride and override
Boolean isOverride is to check if the task has been manually moved, if
so, it will differ the weight to the weight of the two tasks around it,
rather than based off its own weight. to revert manual order, we can
simply recharge the boolean isOverride to false once again and it will
keep its previous place in line if ever wanting to reenable it.